requirements.txt
streamlit
plotly.express
pandas
openpyxl